pyproject.toml
[project]
name = "mash-crew"
version = "0.1.0"
description = "Virtual crew of role-based agents built on top of mash."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Proprietary" }
authors = [{ name = "imsid" }]
dependencies = [
  "mashpy>=0.1.5",
  "python-dotenv>=1.0",
  "google-auth>=2",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "build>=1.2.2",
  "pytest>=8.3.0",
  "twine>=6.1.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["crew*"]

[tool.pylint."messages control"]
disable = [
  "missing-module-docstring",
  "missing-class-docstring",
  "missing-function-docstring",
  "broad-exception-caught"
]
